requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.32.3
sendgrid>=6.11.0
markdown2>=2.5.0 
feedparser>=6.0.11 